requirements.txt
# requirements.txt

# LLM
openai

# Env var
python-dotenv

# NLP
keybert==0.9.0
sentence-transformers==2.7.0
transformers==4.57.6
konlpy

# Web
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# DB
supabase==2.15.2

# Test
pytest==8.3.5
httpx==0.28.1
pytest-mock==3.14.0
